@pettermahlen 2017-01-05
I think you misread the JMH output: 
`21734.998 ±(99.9%) 802.984 ns/op` vs `19520.580 ±(99.9%) 381.775 ns/op` - the confidence interval is smaller/halved in that measurement, but the performance improvement is much smaller (21.7k to 19.5k). I'm in fact a little surprised that you get a performance improvement at all, given that the `. size()` calls might be inlineable. But I suppose it's possible for the size of the list to change outside of this method, so it's probably a good change. 

I would be interested in seeing the JMH test you were using, as well as if there are any benefits from the changes you made to the `.thenApply` lambda (as in, comparing original/optimise array creation/optimise list creation/optimise both). That array size should be an inlineable constant, so it would be a bit surprising to me if that change has any impact.