@dflemstr 2017-02-11
We are running the benchmark with `java.util.Collections$CopiesList` which has some quirks since it is a virtual list of the same element (`final int n; final E element;`).  Thus in the benchmark we are mostly comparing the differences due to the array allocations and it's hard to reason about the list being inlined.

I think the benchmark should be updated to use `ArrayList` with distinct instances of `CompletableFuture`s, and additionally those futures should be reset occasionally so that the benchmark covers not-yet-completed futures.

Of course this is a micro benchmark so the differences in performance are minuscule.  The point of the benchmark was to prove that the old implementation (using streams) was extremely slow in comparison.